pyproject.toml
[project]
name = "ai-helper-itmo"
version = "0.1.0"
description = "Telegram bot for ITMO master programs"
authors = [
    {name = "n.s.gudkov03@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "3.12.7"
dependencies = [
    "python-telegram-bot (>=22.3,<23.0)",
    "requests (>=2.32.4,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "lxml (>=6.0.0,<7.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
